Allow setting of the timeout
Allow setting of the timeout
Setting the XHR timeout (in milliseconds) can only be done after the `open()` method in versions of Internet Explorer. Thus, the workaround approach to creating your own XHR to pass in for `opts.xmlHttpRequest` doesn't work since `open` is called afterwards.

This is also cleaner than creating a custom XHR for each time you'd like to `xr.get` with a more aggressive timeout.